+++ src/components/sections/Header.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useLanguage } from "../contexts/LanguageContext";
+import { useLanguage } from "../../contexts/LanguageContext";
 
 export const Header = ({ data }) => {
   const { language } = useLanguage();
